Update CockroachDB to v20.2.6
This commit updates the version of CockroachDB to v20.2.6. Cluster and
session settings for the CockroachDB adapter have also been updated so
that they are compatible with the new version. Specifically:

  - Cluster setting `sql.distsql.merge_joins.enabled` was removed in
    20.2.
  - Cluster setting `sql.parallel_scans.enabled` was removed in 20.2.
  - Session setting `experimental_serial_normalization` was renamed to
    `serial_normalization` in 20.2.
  - Session setting `vectorize` no longer accepts `auto` as of 20.2.

Additionally, this commit sets the session setting `sql_safe_updates` to
`off` for every generated session. When this is on, mutations without
`WHERE` clauses caused errors. I don't think there is any reason to
keep this setting enabled in the context of SQLancer.